fixed some type bugs

+++ Backend/src/main/java/com/mischievous/fairies/auth/filter/AuthenticatedUser.java
@@ -1,4 +1,4 @@
 package com.mischievous.fairies.auth.filter;
 
-public record AuthenticatedUser(Integer userId, String email) {
+public record AuthenticatedUser(Long userId, String email) {
 }
